feat: encrypt secrets using sops, age & yubikey
Using unstable versions of sops, age and age-plugin-yubikey to handle
secrets and key pairs.

During the activation stage of nixos-rebuild, an age standard
access key is used to decrypt the secrets and sops-nix[^1]
is copying them to the /nix/store.

The yubikey plugin[^2] for age[^3] let us generate and export
an identity, stored on the YubiKey, under the PIV section:

	> ykman piv change-pin
	> ykman piv change-puk
	> ykman piv access change-management-key --generate --protect
	> age-plugin-yubikey --generate
	> age-plugin-yubikey --identity > \
	~/.config/sops/age/id_jonatan.txt
	> age-plugin-yubikey --list
	> ykman piv info

The yubikey identity will contain a reference to the hardware key and
can be stored openly.

The identity of the age standard access key contains a private
key/secret. We should encrypt it with the YubiKey and a strong password
as a backup:

	> cd ~/.config/sops/age/
	> age-keygen -o id_local.txt
	> age-keygen -y -o r_local.txt id_local.txt
	> cat id_jonatan.txt >> keys.txt
	> cat id_local.txt >> keys.txt
	> # yank or copy YubiKey public key manually to r_jonatan.txt
	> cat r_local.txt >> recipients.txt
	> cat r_jonatan.txt >> recipients.txt
	> age -R recipients.txt id_local.txt > id_local.age
	> age -R recipients.txt keys.txt > keys.age
	> rm id_local.txt
	> rm keys.txt
	> # Before development session
	> age -d -i id_jonatan.txt -o keys.txt keys.age
	> # Before build
	> # Uncomment the YubiKey identity in keys.txt. sops-nix will not
	> # interact with the user and fail the build otherwise.
	> # After development session
	> rm keys.txt

sops configuration for the nixos-orion-7000 host:

- Setting sshKeyPaths to an empty array, for both age and gnupg,
because we do not want the tools to assume the keys need to be
copied or used.
- Declaring test secret
- keys.txt in the home age configuration folder is used to lookup age
identities.

After keys in .sops.yaml changed, we re-encrypt the secret files:

	> sops updatekeys secrets/nixos-orion-7000/secrets.yml

sops[^4] keys and creation rules:

Both the user 'jonatan-yubikey-23839166' and the host 'local' can
decrypt secrets under secrets/nixos-orion-7000:

	> sops secrets/nixos-orion-7000/secrets.yml

To check the secret in the nix store after rebuild:

	> sudo cat /run/secrets/secret1

[^1] https://github.com/Mic92/sops-nix
[^2] https://github.com/str4d/age-plugin-yubikey
[^3] https://github.com/FiloSottile/age
[^4] https://github.com/getsops/sops